seperate v1 and v2 seeds, make groupMode to add default in case its not provided through api
