+++ packages/twenty-server/src/engine/workspace-manager/dev-seeder/core/utils/get-page-layout-widget-data-seeds.util.ts
@@ -9,6 +9,7 @@ import { type ObjectMetadataEntity } from 'src/engine/metadata-modules/object-me
 import { PAGE_LAYOUT_TAB_SEEDS } from 'src/engine/workspace-manager/dev-seeder/core/constants/page-layout-tab-seeds.constant';
 import { PAGE_LAYOUT_WIDGET_SEEDS } from 'src/engine/workspace-manager/dev-seeder/core/constants/page-layout-widget-seeds.constant';
 import { generateSeedId } from 'src/engine/workspace-manager/dev-seeder/core/utils/generate-seed-id.util';
+import { getPageLayoutWidgetDataSeedsV2 } from 'src/engine/workspace-manager/dev-seeder/core/utils/get-page-layout-widget-data-seeds-v2.util';
 import { STANDARD_OBJECT_IDS } from 'src/engine/workspace-manager/workspace-sync-metadata/constants/standard-object-ids';
 
 type PageLayoutWidgetDataSeed = {
@@ -36,6 +37,7 @@ const getFieldId = (
 export const getPageLayoutWidgetDataSeeds = (
   workspaceId: string,
   objectMetadataItems: ObjectMetadataEntity[],
+  isDashboardV2Enabled: boolean,
 ): PageLayoutWidgetDataSeed[] => {
   const opportunityObject = objectMetadataItems.find(
     (obj) => obj.standardId === STANDARD_OBJECT_IDS.opportunity,
@@ -61,16 +63,13 @@ export const getPageLayoutWidgetDataSeeds = (
   const opportunityStageFieldId = getFieldId(opportunityObject, 'stage');
 
   const companyIdFieldId = getFieldId(companyObject, 'id');
-  const companyCreatedAtFieldId = getFieldId(companyObject, 'createdAt');
   const companyEmployeesFieldId = getFieldId(companyObject, 'employees');
   const companyArrFieldId = getFieldId(companyObject, 'annualRecurringRevenue');
-  const companyNameFieldId = getFieldId(companyObject, 'name');
   const companyLinkedinLinkFieldId = getFieldId(companyObject, 'linkedinLink');
   const companyAddressFieldId = getFieldId(companyObject, 'address');
 
   const personIdFieldId = getFieldId(personObject, 'id');
   const personCityFieldId = getFieldId(personObject, 'city');
-  const personJobTitleFieldId = getFieldId(personObject, 'jobTitle');
 
   const opportunityIdFieldId = getFieldId(opportunityObject, 'id');
 
@@ -79,7 +78,7 @@ export const getPageLayoutWidgetDataSeeds = (
   const rocketIdFieldId = getFieldId(rocketObject, 'id');
   const rocketCreatedAtFieldId = getFieldId(rocketObject, 'createdAt');
 
-  return [
+  const v1Widgets = [
     // Sales Overview Tab Widgets
     {
       id: generateSeedId(
@@ -128,35 +127,6 @@ export const getPageLayoutWidgetDataSeeds = (
           }
         : null,
       objectMetadataId: rocketObject?.id ?? null,
-    },
-    {
-      id: generateSeedId(
-        workspaceId,
-        PAGE_LAYOUT_WIDGET_SEEDS.SALES_REVENUE_FORECAST,
-      ),
-      pageLayoutTabId: generateSeedId(
-        workspaceId,
-        PAGE_LAYOUT_TAB_SEEDS.SALES_OVERVIEW,
-      ),
-      title: 'Revenue Forecast',
-      type: WidgetType.GRAPH,
-      gridPosition: { row: 0, column: 7, rowSpan: 8, columnSpan: 5 },
-      configuration:
-        isDefined(opportunityAmountFieldId) &&
-        isDefined(opportunityCloseDateFieldId)
-          ? {
-              graphType: 'LINE',
-              aggregateFieldMetadataId: opportunityAmountFieldId,
-              aggregateOperation: AggregateOperations.SUM,
-              primaryAxisGroupByFieldMetadataId: opportunityCloseDateFieldId,
-              primaryAxisOrderBy: 'FIELD_ASC',
-              axisNameDisplay: AxisNameDisplay.NONE,
-              displayDataLabel: false,
-              timezone: 'UTC',
-              firstDayOfTheWeek: CalendarStartDay.MONDAY,
-            }
-          : null,
-      objectMetadataId: opportunityObject?.id ?? null,
     },
     {
       id: generateSeedId(
@@ -274,34 +244,6 @@ export const getPageLayoutWidgetDataSeeds = (
     {
       id: generateSeedId(
         workspaceId,
-        PAGE_LAYOUT_WIDGET_SEEDS.CUSTOMER_NEW_OVER_TIME,
-      ),
-      pageLayoutTabId: generateSeedId(
-        workspaceId,
-        PAGE_LAYOUT_TAB_SEEDS.CUSTOMER_OVERVIEW,
-      ),
-      title: 'New Customers Over Time',
-      type: WidgetType.GRAPH,
-      gridPosition: { row: 0, column: 3, rowSpan: 6, columnSpan: 5 },
-      configuration:
-        isDefined(companyIdFieldId) && isDefined(companyCreatedAtFieldId)
-          ? {
-              graphType: 'LINE',
-              aggregateFieldMetadataId: companyIdFieldId,
-              aggregateOperation: AggregateOperations.COUNT,
-              primaryAxisGroupByFieldMetadataId: companyCreatedAtFieldId,
-              primaryAxisOrderBy: 'FIELD_ASC',
-              axisNameDisplay: AxisNameDisplay.NONE,
-              displayDataLabel: false,
-              timezone: 'UTC',
-              firstDayOfTheWeek: CalendarStartDay.MONDAY,
-            }
-          : null,
-      objectMetadataId: companyObject?.id ?? null,
-    },
-    {
-      id: generateSeedId(
-        workspaceId,
         PAGE_LAYOUT_WIDGET_SEEDS.CUSTOMER_COMPANIES_BY_SIZE,
       ),
       pageLayoutTabId: generateSeedId(
@@ -362,57 +304,6 @@ export const getPageLayoutWidgetDataSeeds = (
     {
       id: generateSeedId(
         workspaceId,
-        PAGE_LAYOUT_WIDGET_SEEDS.CUSTOMER_REVENUE_DISTRIBUTION,
-      ),
-      pageLayoutTabId: generateSeedId(
-        workspaceId,
-        PAGE_LAYOUT_TAB_SEEDS.CUSTOMER_ANALYTICS,
-      ),
-      title: 'Revenue Distribution',
-      type: WidgetType.GRAPH,
-      gridPosition: { row: 0, column: 4, rowSpan: 2, columnSpan: 3 },
-      configuration:
-        isDefined(companyArrFieldId) && isDefined(companyNameFieldId)
-          ? {
-              graphType: 'PIE',
-              aggregateFieldMetadataId: companyArrFieldId,
-              aggregateOperation: AggregateOperations.SUM,
-              groupByFieldMetadataId: companyNameFieldId,
-              orderBy: 'VALUE_DESC',
-              displayDataLabel: true,
-              timezone: 'UTC',
-              firstDayOfTheWeek: CalendarStartDay.MONDAY,
-            }
-          : null,
-      objectMetadataId: companyObject?.id ?? null,
-    },
-    {
-      id: generateSeedId(
-        workspaceId,
-        PAGE_LAYOUT_WIDGET_SEEDS.CUSTOMER_AVERAGE_ARR,
-      ),
-      pageLayoutTabId: generateSeedId(
-        workspaceId,
-        PAGE_LAYOUT_TAB_SEEDS.CUSTOMER_ANALYTICS,
-      ),
-      title: 'Average ARR',
-      type: WidgetType.GRAPH,
-      gridPosition: { row: 0, column: 7, rowSpan: 6, columnSpan: 5 },
-      configuration: isDefined(companyArrFieldId)
-        ? {
-            graphType: 'GAUGE',
-            aggregateFieldMetadataId: companyArrFieldId,
-            aggregateOperation: AggregateOperations.AVG,
-            displayDataLabel: true,
-            timezone: 'UTC',
-            firstDayOfTheWeek: CalendarStartDay.MONDAY,
-          }
-        : null,
-      objectMetadataId: companyObject?.id ?? null,
-    },
-    {
-      id: generateSeedId(
-        workspaceId,
         PAGE_LAYOUT_WIDGET_SEEDS.CUSTOMER_LINKEDIN_COUNT,
       ),
       pageLayoutTabId: generateSeedId(
@@ -432,33 +323,6 @@ export const getPageLayoutWidgetDataSeeds = (
             firstDayOfTheWeek: CalendarStartDay.MONDAY,
           }
         : null,
-      objectMetadataId: companyObject?.id ?? null,
-    },
-    {
-      id: generateSeedId(
-        workspaceId,
-        PAGE_LAYOUT_WIDGET_SEEDS.CUSTOMER_LINKEDIN_DISTRIBUTION,
-      ),
-      pageLayoutTabId: generateSeedId(
-        workspaceId,
-        PAGE_LAYOUT_TAB_SEEDS.CUSTOMER_OVERVIEW,
-      ),
-      title: 'Companies by LinkedIn (Field Permission Test)',
-      type: WidgetType.GRAPH,
-      gridPosition: { row: 6, column: 0, rowSpan: 4, columnSpan: 6 },
-      configuration:
-        isDefined(companyIdFieldId) && isDefined(companyLinkedinLinkFieldId)
-          ? {
-              graphType: 'PIE',
-              aggregateFieldMetadataId: companyIdFieldId,
-              aggregateOperation: AggregateOperations.COUNT,
-              groupByFieldMetadataId: companyLinkedinLinkFieldId,
-              orderBy: 'VALUE_DESC',
-              displayDataLabel: true,
-              timezone: 'UTC',
-              firstDayOfTheWeek: CalendarStartDay.MONDAY,
-            }
-          : null,
       objectMetadataId: companyObject?.id ?? null,
     },
 
@@ -516,33 +380,6 @@ export const getPageLayoutWidgetDataSeeds = (
 
     // Team Metrics Tab Widgets
     {
-      id: generateSeedId(
-        workspaceId,
-        PAGE_LAYOUT_WIDGET_SEEDS.TEAM_CONTACT_ROLES,
-      ),
-      pageLayoutTabId: generateSeedId(
-        workspaceId,
-        PAGE_LAYOUT_TAB_SEEDS.TEAM_METRICS,
-      ),
-      title: 'Contact Roles',
-      type: WidgetType.GRAPH,
-      gridPosition: { row: 0, column: 0, rowSpan: 4, columnSpan: 6 },
-      configuration:
-        isDefined(personIdFieldId) && isDefined(personJobTitleFieldId)
-          ? {
-              graphType: 'PIE',
-              aggregateFieldMetadataId: personIdFieldId,
-              aggregateOperation: AggregateOperations.COUNT,
-              groupByFieldMetadataId: personJobTitleFieldId,
-              orderBy: 'VALUE_DESC',
-              displayDataLabel: true,
-              timezone: 'UTC',
-              firstDayOfTheWeek: CalendarStartDay.MONDAY,
-            }
-          : null,
-      objectMetadataId: personObject?.id ?? null,
-    },
-    {
       id: generateSeedId(workspaceId, PAGE_LAYOUT_WIDGET_SEEDS.TEAM_OPEN_TASKS),
       pageLayoutTabId: generateSeedId(
         workspaceId,
@@ -563,5 +400,12 @@ export const getPageLayoutWidgetDataSeeds = (
         : null,
       objectMetadataId: taskObject?.id ?? null,
     },
+  ];
+
+  return [
+    ...v1Widgets,
+    ...(isDashboardV2Enabled
+      ? getPageLayoutWidgetDataSeedsV2(workspaceId, objectMetadataItems)
+      : []),
   ];
 };
