+++ packages/twenty-server/src/engine/workspace-manager/dev-seeder/core/utils/seed-page-layout-widgets.util.ts
@@ -20,6 +20,7 @@ export const seedPageLayoutWidgets = async ({
   const pageLayoutWidgets = getPageLayoutWidgetDataSeeds(
     workspaceId,
     objectMetadataItems,
+    isDashboardV2Enabled,
   ).map((widget) => {
     const validatedConfiguration = widget.configuration
       ? validateAndTransformWidgetConfiguration({
